+++ HISTORY.md
@@ -1,5 +1,9 @@
 # HEAD
 
+* Remove `Chronic::InvalidArgumentException` and instead use
+  `ArgumentError`
+* Remove `Time.construct` and use `Chronic.construct` in place of this
+* Remove `Time#to_minidate`, instead use `Chronic::MiniDate.from_time(time)`
 * Add support for handling generic timestamp for Ruby 1.9+
 
 # 0.5.0 / 2011-07-01